+++ packages/stream_chat/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Upcoming
+## 9.22.0
 
 ✅ Added